+++ models/chgnet/ctk_structure_viewer.py
@@ -3,7 +3,7 @@ from crystal_toolkit.helpers.utils import hook_up_fig_with_struct_viewer
 from pymatviz.enums import Key
 from pymatviz.utils import PLOTLY
 
-from matbench_discovery.preds import PRED_FILES
+from matbench_discovery.preds import PredFiles
 
 __author__ = "Janosh Riebesell"
 __date__ = "2023-03-07"
@@ -18,14 +18,14 @@ Then open http://localhost:8000 in your browser.
 e_form_2000 = "e_form_per_atom_chgnet_2000"
 e_form_500 = "e_form_per_atom_chgnet_500"
 
-df_chgnet = pd.read_json(PRED_FILES.CHGNet.replace(".csv.gz", ".json.gz"))
+df_chgnet = pd.read_json(PredFiles.chgnet.path.replace(".csv.gz", ".json.gz"))
 df_chgnet = df_chgnet.set_index(Key.mat_id)
 
-df_chgnet_2000 = pd.read_csv(PRED_FILES.CHGNet)
+df_chgnet_2000 = pd.read_csv(PredFiles.chgnet.path)
 df_chgnet_2000 = df_chgnet_2000.set_index(Key.mat_id).add_suffix("_2000")
 df_chgnet[list(df_chgnet_2000)] = df_chgnet_2000
 
-df_chgnet_500 = pd.read_csv(PRED_FILES.CHGNet.replace("-06", "-04"))
+df_chgnet_500 = pd.read_csv(PredFiles.chgnet.path.replace("-06", "-04"))
 df_chgnet_500 = df_chgnet_500.set_index(Key.mat_id).add_suffix("_500")
 df_chgnet[list(df_chgnet_500)] = df_chgnet_500