+++ models/chgnet/test_chgnet.py
@@ -20,7 +20,7 @@ from pymatviz.enums import Key
 from tqdm import tqdm
 
 from matbench_discovery import timestamp, today
-from matbench_discovery.data import DATA_FILES, as_dict_handler, df_wbm
+from matbench_discovery.data import DataFiles, as_dict_handler, df_wbm
 from matbench_discovery.enums import MbdKey, Task
 from matbench_discovery.plots import wandb_scatter
 from matbench_discovery.slurm import slurm_submit
@@ -59,8 +59,8 @@ if os.path.isfile(out_path):
 
 # %%
 data_path = {
-    Task.RS2RE: DATA_FILES.wbm_computed_structure_entries,
-    Task.IS2RE: DATA_FILES.wbm_initial_structures,
+    Task.RS2RE: DataFiles.wbm_computed_structure_entries.path,
+    Task.IS2RE: DataFiles.wbm_initial_structures.path,
 }[task_type]
 print(f"\nJob started running {timestamp}")
 print(f"{data_path=}")